+++ src/BaseDatasource/Contracts/BaseDatasourceContract.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace ForestAdmin\AgentPHP\BaseDatasource\Contracts;
+
+use ForestAdmin\AgentPHP\DatasourceToolkit\Components\Contracts\DatasourceContract;
+
+interface BaseDatasourceContract extends DatasourceContract
+{
+}
