+++ src/DatasourceDummy/Collections/Book.php
@@ -8,7 +8,7 @@ use ForestAdmin\AgentPHP\DatasourceToolkit\Components\Query\ConditionTree\Operat
 use ForestAdmin\AgentPHP\DatasourceToolkit\Schema\ColumnSchema;
 use ForestAdmin\AgentPHP\DatasourceToolkit\Schema\Concerns\PrimitiveType;
 
-class Book extends BaseCollection
+class Book extends DummyCollection
 {
     protected $records = [
         [
